+++ src/api/api.js
@@ -1,11 +1,12 @@
 import { getConfigValue } from '../config.js';
+import {
+  EmptyResponseError,
+  InternalServerError,
+  NotAuthorizedError,
+  PuzzleNotFoundError,
+} from '../errors/apiErrors.js';
 import { sizeOfStringInKb } from '../formatting.js';
 import { logger } from '../logger.js';
-import {
-  extractTextContentOfMain,
-  parseResponseMessage,
-  sanitizeMessage,
-} from './parseSubmissionResponse.js';
 import { puzzleAnswerUrl, puzzleInputUrl } from './urls.js';
 /**
  * Creates a headers object which can be passed to fetch.
@@ -23,46 +24,33 @@ const getHeaders = (authenticationToken) => ({
  * @param {Number} day - The day of the puzzle.
  * @param {Promise<String>} authenticationToken - Token to authenticate with aoc.
  */
-export const downloadInput = async (year, day, authenticationToken) => {
+export const getInput = async (year, day, authenticationToken) => {
   logger.debug('downloading input file for year: %s, day: %s', year, day);
-
   if (!authenticationToken) {
-    throw new Error(
-      'Authentication Token is required to query advent of code.'
-    );
+    throw new NotAuthorizedError();
   }
 
-  // query api
   const url = puzzleInputUrl(year, day);
   logger.debug('querying url for input: %s', url);
   const response = await fetch(url, {
     headers: getHeaders(authenticationToken),
   });
 
-  // bad request, authentication failed.
   if (response.status === 400) {
-    throw new Error('Authentication failed, double check authentication token');
+    throw new NotAuthorizedError();
   }
-  // not found, invalid day or year.
   if (response.status === 404) {
-    throw new Error('That year/day combination could not be found');
+    throw new PuzzleNotFoundError(year, day);
   }
-  // handle all other error status codes
   if (!response.ok) {
-    throw new Error(
-      `Unexpected server error while downloading input file, error: ${response.status} - ${response.statusText}`
-    );
+    throw new InternalServerError(response.status, response.statusText);
   }
 
-  // expect text of response is the input.
   const text = await response.text();
-
-  if (!text) {
-    throw new Error('Advent of code returned empty input');
-  }
-
   logger.debug('downloaded: %s', sizeOfStringInKb(text));
-
+  if (!text) {
+    throw new EmptyResponseError();
+  }
   return text;
 };
 
@@ -74,22 +62,19 @@ export const downloadInput = async (year, day, authenticationToken) => {
  * @param {String|Number} solution - The solution to test.
  * @param {String} authenticationToken - Token to authenticate with aoc.
  */
-export const submitSolution = async (
+export const postAnswer = async (
   year,
   day,
   level,
   solution,
   authenticationToken
 ) => {
-  logger.debug('submitting solution to advent of code', { year, day, level });
+  logger.debug('posting answer to advent of code', { year, day, level });
 
   if (!authenticationToken) {
-    throw new Error(
-      'Authentication Token is required to query advent of code.'
-    );
+    throw new NotAuthorizedError();
   }
 
-  // post to api
   const url = puzzleAnswerUrl(year, day);
   logger.debug('posting to url: %s', url);
   const response = await fetch(url, {
@@ -101,34 +86,20 @@ export const submitSolution = async (
     body: `level=${level}&answer=${solution}`,
   });
 
-  // bad request, authentication failed.
-  // as of writing advent returns a 302 to redirect the user to the puzzle page on fail
-  // but check 400 too just incase.
+  // server can sometimes redirect (302) back to puzzle page if auth fails.
   if (response.status === 400 || response.status === 302) {
-    throw new Error('Authentication failed, double check authentication token');
+    throw new NotAuthorizedError();
   }
-  // not found, invalid day or year.
   if (response.status === 404) {
-    throw new Error('That year/day combination could not be found');
+    throw new PuzzleNotFoundError(year, day);
   }
-  // bail on any other type of http error
   if (!response.ok) {
-    throw new Error(
-      `Unexpected server error while posting solution, error: ${response.status} - ${response.statusText}`
-    );
+    throw new InternalServerError(response.status, response.statusText);
   }
 
-  // advent of code doesn't return status codes, we have to parse the html.
-  // grab the text content of the <main> element which contains the message we need.
-  const responseMessage = sanitizeMessage(
-    extractTextContentOfMain(await response.text())
-  );
-
-  if (!responseMessage) {
-    throw new Error('Unable get message from advent of code response.');
+  const text = await response.text();
+  if (!text) {
+    throw new EmptyResponseError();
   }
-
-  // the content of the message tells us what happened
-  // parse this message to determine the submission result.
-  return parseResponseMessage(responseMessage);
+  return text;
 };