+++ src/api/index.js
@@ -8,17 +8,17 @@ import { logger } from '../logger.js';
  * Decorates the api methods with rate limiting.
  * @param {Object} module - The imported api to decorate.
  */
-const decorateApiWithRateLimiting = ({ downloadInput, submitSolution }) => {
+const decorateApiWithRateLimiting = ({ getInput, postAnswer }) => {
   logger.debug('decorating api with rate limiting');
   return {
-    downloadInput: rateLimitDecorator(
-      downloadInput,
-      rateLimitedActions.downloadInput,
+    getInput: rateLimitDecorator(
+      getInput,
+      rateLimitedActions.getInput,
       'Timeout period for downloading an input file has not expired.'
     ),
-    submitSolution: rateLimitDecorator(
-      submitSolution,
-      rateLimitedActions.submitAnswer,
+    postAnswer: rateLimitDecorator(
+      postAnswer,
+      rateLimitedActions.postAnswer,
       'Timeout period for submitting an answer has not expired.'
     ),
   };
@@ -26,8 +26,8 @@ const decorateApiWithRateLimiting = ({ downloadInput, submitSolution }) => {
 
 const useMockApi = getConfigValue('aoc.mockApi.enabled');
 
-const { downloadInput, submitSolution } = useMockApi
+const { getInput, postAnswer } = useMockApi
   ? await import('./mockApi.js')
   : decorateApiWithRateLimiting(await import('./api.js'));
 
-export { downloadInput, submitSolution };
+export { getInput, postAnswer };