+++ src/api/mockApi.js
@@ -10,7 +10,7 @@ import { logger } from '../logger.js';
  * @param {String} authenticationToken - Token to authenticate with aoc.
  * @returns {Promise<string>}
  */
-export const downloadInput = async (year, day, authenticationToken) => {
+export const getInput = async (year, day, authenticationToken) => {
   logger.debug('downloading mock input file for year: %s, day: %s', year, day);
   return `${[...Array(100).keys()].map((x) => x + 1).join('\n')}\n`;
 };
@@ -23,19 +23,15 @@ export const downloadInput = async (year, day, authenticationToken) => {
  * @param {String|Number} solution - The solution to test.
  * @param {String} authenticationToken - Token to authenticate with aoc.
  */
-export const submitSolution = async (
+export const postAnswer = async (
   year,
   day,
   level,
   solution,
   authenticationToken
 ) => {
-  logger.debug('submitting answer to mock api');
-  const correct = getConfigValue('aoc.mockApi.answerCorrect');
-  return {
-    correct,
-    message: correct
-      ? 'Great Job you answered correct!'
-      : "That's the wrong answer!",
-  };
+  logger.debug('posting answer to mock api');
+  return getConfigValue('aoc.mockApi.answerCorrect')
+    ? "that's the right answer"
+    : "that's not the right answer";
 };