+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix submission response parsing. Now handles more server responses with better output to the user. ([#223](https://github.com/beakerandjake/advent-of-code-runner/issues/223))
+### Changed
+- Remove dependencies:  dom-serializer, domutils, htmlparser2
 
 ## [1.5.0] - 2023-11-16
 ### Added